+++ test/biotcm/apps/test_pubmed_gene_miner.rb
@@ -8,12 +8,14 @@ describe BioTCM::Apps::PubmedGeneMiner do
   it 'mine genes online' do
     res = @miner.mine_online('(IBD[Title/Abstract]) OR (CRC[Title/Abstract]) AND Cancer[Title/Abstract]')
     assert_instance_of(Hash, res)
-    assert(res['IL2'] > 0)
+    assert_instance_of(Array, res['IL2'])
+    refute_empty(res['IL2'])
   end
 
   it 'mine genes offline' do
     res = @miner.mine_offline('(IBD[Title/Abstract]) OR (CRC[Title/Abstract]) AND Cancer[Title/Abstract] AND Obesity[Title/Abstract] AND IL-2[Title/Abstract]')
     assert_instance_of(Hash, res)
-    assert(res['IL2'] > 0)
+    assert_instance_of(Array, res['IL2'])
+    refute_empty(res['IL2'])
   end
 end